Split connection processing part of loop out
This is so that implementors can subclass the server to do things other
than just handle websockets. For example, you might want a GameServer
that processes server-side events at regular ticks. This is hard to do
if the entire main loop of the server is in a single scope: you'd have
to override the entire run() method.

This commit simply moves the contents of the inside of the main loop in
run() into a separate method (processSockets()). It's understood this
method might take upwards of five seconds to return.