Removed check for false values on connection->send
This was preventing one from sending any of the following:

0, "0", "" and NULL (which is recieved as an empty string in JS clients)

The use case for this is if you want to return a count of something - 0
needs to be an option.

The use case for empty strings and NULL is to save confusion when a method is not
returning what it is expected to - having the websocket server send
nothing at all in such cases makes them difficult to debug.